Split probe env key + extension into their own files (Aziz policy)
One major type per Swift file, matching the MinimalModeInvalidationProbe
three-file layout exactly. No content changes.
