+++ Sources/Debug/SidebarLazyContractProbeKey.swift
@@ -1,0 +1,7 @@
+import SwiftUI
+
+#if DEBUG
+struct SidebarLazyContractProbeKey: EnvironmentKey {
+    static let defaultValue = SidebarLazyContractProbe()
+}
+#endif
